requirements.txt
asyncio==3.4.3
asyncpg==0.24.0
PyYAML==5.4.1
